Update move settings to own page
